Move fallbackhumanizer tests
