+++ src/libs/humanizer/modules/FallbackHumanizer/index.ts
@@ -1,0 +1,3 @@
+import { fallbackHumanizer } from './fallBackHumanizer'
+
+export default fallbackHumanizer
